+++ database/migrations/2021_05_04_032450_create_doctors_table.php
@@ -18,11 +18,11 @@ class CreateDoctorsTable extends Migration
             $table->string('name');
             $table->string('dob');
             $table->string('prof_name');
-            $table->unsignedInteger('id_profession');
-            $table->unsignedInteger('id_branch');
+            $table->unsignedBigInteger('id_profession');
+            $table->unsignedBigInteger('id_branch');
             $table->Integer('status');
             $table->string('cabinet');
-            $table->unsignedInteger('id_user');
+            $table->unsignedBigInteger('id_user');
             $table->string('work_day')->default('1,2,3,4,5');
 
             $table->foreign('id_profession')->references('id')->on('professions')->onUpdate('CASCADE');
